Divide into task and commands packages
