+++ src/main/java/duke/Storage.java
@@ -1,4 +1,9 @@
 package duke;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
+
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
