+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,17 +1,22 @@
-package duke;
+package duke.commands;
 
-public class UnmarkCommand extends Command {
-    public String input;
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.commands.Command;
 
-    public UnmarkCommand(String input) {
+public class DeleteCommand extends Command {
+    private final String input;
+
+    public DeleteCommand(String input) {
         this.input = input;
     }
 
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
         try {
             int index = Integer.parseInt(input.trim()) - 1;
-            taskList.get(index).markAsUndone();
-            ui.showUnmarkedTask(taskList.get(index));
+            ui.showDeleteTask(taskList.deleteTask(index), taskList);
         } catch (NumberFormatException e) {
             throw new DukeException("Input a valid number!");
         }
