+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,4 +1,10 @@
-package duke;
+package duke.commands;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.commands.Command;
 
 public class ExitCommand extends Command {
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
