+++ src/main/java/duke/commands/GreetCommand.java
@@ -1,4 +1,10 @@
-package duke;
+package duke.commands;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.commands.Command;
 
 public class GreetCommand extends Command {
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
