+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,4 +1,10 @@
-package duke;
+package duke.commands;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.commands.Command;
 
 public class MarkCommand extends Command {
     public String input;
