+++ src/main/java/duke/commands/OnCommand.java
@@ -1,4 +1,10 @@
-package duke;
+package duke.commands;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.commands.Command;
 
 public class OnCommand extends Command {
     public String input;
