+++ src/main/java/duke/task/ToDo.java
@@ -1,5 +1,4 @@
-package duke;
-import java.time.LocalDate;
+package duke.task;
 
 public class ToDo extends Task {
     public ToDo(String description) {
